refactor: copy repo processing functions to new process_repo.py module
